Temp remote stuff
found something for the remote. i got a quite short "deadline", so let's try this.

flashing will be done using the arduino uno board without the processor, the nonnector on the remote and a bit of "engineering" to connect to pd0 and pd1 pins. in theory this will then be connected to a pc with an arduino uno compatible cable. software used for flashing will be the arduino app set to flash an atmega168. other settings will be figured out on site, but are expected to be mostly same as for the arduino uno board.